move kafka node to file
